requirements.txt
adb==1.3.0
aniso8601==2.0.1
arandr==0.1.9
asn1crypto==0.22.0
certifi==2017.11.5
cffi==1.10.0
click==6.7
cryptography==2.0
dnspython==1.15.0
enum34==1.1.6
Flask==0.12.1
Flask-Ask==0.9.6
google-api-python-client==1.6.1
gyp==0.1
httplib2==0.9.2
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
keyring==10.6.0
keyrings.alt==3.0
libusb1==1.6.2
M2Crypto==0.27.0
MarkupSafe==1.0
mercurial==4.5.3
oauth2client==4.0.0
paho-mqtt==1.3.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycairo==1.16.2
pycparser==2.18
pycrypto==2.6.1
pydhcplib==0.6.2
pygobject==3.26.1
pyOpenSSL==17.0.0
python-dateutil==2.6.1
pyxdg==0.25
PyYAML==3.12
requests==2.13.0
rsa==3.4.2
scapy==2.3.1
SecretStorage==2.3.1
simplejson==3.11.1
six==1.10.0
transmissionrpc==0.11
trimage==1.0.2
typing==3.5.3.0
unity-lens-photos==1.0
uritemplate==3.0.0
urllib3==1.22
wakeonlan==0.2.2
Werkzeug==0.12.2
youtube-dl==2017.6.5
